Refacto : Admin

+++ src/Controller/Admin/Comment/DeleteController.php
@@ -1,38 +1,27 @@
 <?php
 
-namespace App\Controller\Admin;
+namespace App\Controller\Admin\Comment;
 
 use App\Entity\Comment;
-use App\Repository\CommentRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
-class CommentsController extends AbstractController
+class DeleteController extends AbstractController
 {
-
   private $em;
-
   public function __construct(EntityManagerInterface $em)
   {
     $this->em = $em;
   }
 
-  #[Route('/admin/comments', name: 'admin_comments')]
-  public function index(CommentRepository $commentRepository)
-  {
-    $comments = $commentRepository->findAll();
-    return $this->render('admin/comment/comments.html.twig', [
-      'current' => 'comments',
-      'comments' => $comments
-    ]);
-  }
-
-
-
   #[Route('/admin/comment/{id}/delete', name: 'admin_comment_delete')]
-  public function delete(Comment $comment)
+  public function delete(Comment $comment = null)
   {
+    if (null === $comment) {
+      throw new NotFoundHttpException('Comment not found !');
+    }
 
     $this->em->remove($comment);
     $this->em->flush();
